src/requirements.txt
azure-functions
requests
datetime
azure-identity
azure-mgmt-resource
azure-functions